refactor(all): minor fixes + refactor
