+++ src/app/modules/main-page/center-box/center-box.component.html
@@ -1,10 +1,10 @@
 <div class="flexbox">
   <div class="top-box">
-    <button mat-button class="add-button" (click)="onAddExpenseClick()">Add an Expense</button>
+    <button mat-button class="add-button" (click)="onAddExpenseClick()" type="button">Add an Expense</button>
     <div class="text">{{displayString}}</div>
 
 
-    <button mat-button class="settle-button">Settle Up</button>
+    <button mat-button class="settle-button" type="button">Settle Up</button>
 
   </div>
   <app-cash-status></app-cash-status>
