+++ src/app/modules/main-page/header/header.component.html
@@ -11,7 +11,7 @@
     <div class="settings-flex">
       <mat-icon class="setting-icon" [matMenuTriggerFor]="menu">settings</mat-icon>
       <mat-menu #menu="matMenu" class="dropdown" xPosition="after">
-        <button mat-menu-item (click)="onLogout()">Logout</button>
+        <button mat-menu-item (click)="onLogout()" type="button">Logout</button>
       </mat-menu>
     </div>
   </div>
