+++ src/app/modules/main-page/side-bar-left/friends/friends.component.html
@@ -24,6 +24,6 @@
   </mat-form-field>
 </div>
 
-<button mat-button (click)="addFriend(newFriend)" class="add-button" type="submit" [disabled]="!newFriend">Add new
+<button mat-button (click)="addFriend(newFriend)" class="add-button" type="button" [disabled]="!newFriend">Add new
   friend
 </button>
