+++ src/app/modules/main-page/side-bar-left/groups/groups.component.html
@@ -9,5 +9,5 @@
     {{group}}</li>
 </ul>
 
-<button mat-button class="add-button" (click)="newGroupLink()">Add new group</button>
+<button mat-button class="add-button" (click)="newGroupLink()" type="button">Add new group</button>
