+++ src/app/modules/main-page/side-bar-left/view-choice/view-choice.component.html
@@ -1,19 +1,19 @@
 <mat-list>
   <mat-list-item>
-    <button mat-button (click)="onClick('Dashboard')">
+    <button mat-button (click)="onClick('Dashboard')" type="button">
       <img src="assets/dashboard.png" class="dashboard-icon"
            alt="dashboard">
       Dashboard
     </button>
   </mat-list-item>
   <mat-list-item>
-    <button mat-button (click)="onClick('Recent Activities')">
+    <button mat-button (click)="onClick('Recent Activities')" type="button">
       <mat-icon>flag</mat-icon>
       Recent Activities
     </button>
   </mat-list-item>
   <mat-list-item>
-    <button mat-button (click)="onClick('All Expenses')">
+    <button mat-button (click)="onClick('All Expenses')" type="button">
       <mat-icon> format_list_bulleted</mat-icon>
       All Expenses
     </button>
